requirements.txt
pdfplumber==0.10.3
openpyxl==3.1.2
Pillow
# cryptography==41.0.7  # Sometimes needed
openpyxl
easyocr==1.7.0
pdf2image
numpy
PyMuPDF